fix: load variant helpers via CJS for Expo app.config
Expo evaluates app.config through CommonJS require, which cannot resolve
@/ aliases or .ts submodules. Centralize naming logic in variant.cjs and
re-export from variant.ts for the app and variant script.
